Add spell-check GitHub Actions workflow
The repository will be checked for commonly misspelled words on every pull request and push.

In the event of false positives, add the word you wish the check to ignore to etc/codespell-ignore-words-list.txt. Always use lowercase, as the ignore words list is case-sensitive and the list of commonly misspelled words is in lowercase.