Add GitHub Actions workflow to run Python unit tests
Run unit tests on every PR and push.

I modified one of the unit tests to make it use the absolute path to the test data rather than the relative path. The reason is to avoid having to add an extra line to the workflow file to cd to the test folder.