Report deltas for all sketches
Previously, the user specified a single sketch used to track size data. This might not provide an accurate indicator of size changes.

In order to make the report easy to read, the summary report shows the range of memory usage change for each FQBN.

Full reports are provided as folded sections of the comment for when detailed information is required. The full reports are available in both markdown table (for browsing in the comment) and CSV (for importing into a spreadsheet program when the markdown table is too large, or custom data sorting is needed) forms.