Add more log output
Previously, the action provided log output only on error, or when verbosity is enabled (which there is no mechanism for the user to do). Adding some additional log output may provide troubleshooting assistance to the user without excessively cluttering the log.

The logging system is now used primarily for debugging during development. The logging output is less user friendly and is incompatible with the GitHub Actions log level commands system.

The prints prepended with ::debug:: are only displayed when the ACTIONS_STEP_DEBUG secret is set to true. So this doesn't cause any additional log content during normal use of the action.
https://github.com/actions/toolkit/blob/master/docs/action-debugging.md#how-to-access-step-debug-logs