Skip reports with mismatched commit hash
The deltas report heading now uses the hash from the report (which is the sensible thing to do, since the report data is for that hash, but the check for an existing report compares the PR head ref provided by the GitHub API to the hash from the report heading (because the check for existing report must be done before retrieving the report to avoid an excessive number of API requests. Without this check, if for some reason the two hashes didn't match, it would result in the deltas report being done over and over again.