Support using sketches reports from local path
The action was previously designed to only run from a scheduled workflow. The
reason is that it needs a token with write permissions to comment on the PR, but
due to security restrictions there is no way to have such a token when a
workflow is triggered by a pull_request event from a fork.

This is problematic for private repos because if the schedule is set to a short
interval the action will use up the free GitHub actoins minutes allocation
quickly (public repos have unlimited minutes). If the schedule is set to a long
interval, then there is a long potential wait time for the report.

A common work flow in private repos is for PRs to be submitted from branches,
not forks, which makes it possible to trigger the action from the PR.

Running from a pull request triggered workflow should actually work as the
action was, but the method of finding the artifact is very inefficient and
unintuitive in that context.

Recently, GitHub added the ability for private repositories to allow write
permissions for workflows triggered by pull requests, making it even more likely
this method of using the action will be found useful:
https://docs.github.com/en/github/administering-a-repository/disabling-or-limiting-github-actions-for-a-repository#enabling-workflows-for-private-repository-forks