Use appropriate default value for sketches-reports-source input
In the time since it was first conceived, the use of the report file created by the compile-sketches action has been expanded from only being a way to pass memory usage change data to the report-size-deltas action to a general purpose report of data generated from the sketch compilations, including:

- Non-deltas size data consumed by `arduino/report-size-trends`
- Compilation status for each sketch/board combination
- Compiler warning count

It's possible additional information will be added to the report over time.

For this reason, the previous default of the `arduino/compile-sketches` action's `sketches-report-path` input: `size-deltas-reports` was no longer appropriate and had to be changed to avoid confusion. Even though `arduino/report-size-deltas` only uses the deltas data from the report, and so the previous default was somewhat logical, I think it's best for the same default name to be used for the folder and workflow artifacts containing the reports consistently across all the actions that use this report.

Unfortunately, this will be a breaking change for people who have not specified the action's `sketches-reports-source` input in their workflow and are thus relying on the default value.