Skip expired workflow artifacts (#10)
Workflow artifacts expire after a default of 90 days. For some inscrutable reason, the GitHub workflow artifacts API
still lists these artifacts, even though they are no longer available. Previously, if the repository contained an open
PR with an expired artifact, it caused the action to fail when it tried to download the non-existent artifact and got an
HTTP 500 error.