Set up basic integration tests
Although the action's Python code itself is currently well tested via unit tests, this does not provide any checks for
how the action interacts with the GitHub Actions framework and GitHub API.

Even though it won't provide full verification of the report, simply running the action provides a
"smoke test" as well as facilitating manual verification.

On every pull request event, the action will run and source the sketches reports test data from a local path.

When the job is triggered by an event from a fork, it will fail due to the access token not having the write permissions
required to comment on the PR thread. The workflow is triggered to still pass under these conditions. Checking the log to
see whether the action failed in the expected manner may still provide a useful validation of the PR.

When the PR is from a branch of the repository, the deltas report will be commented on the PR thread.

The workflow is also configured to run on push, schedule, workflow_dispatch, and repository_dispatch events. In this case
it sources the sketches report from a workflow artifact. A dummy PR will be left permanently open in the repository with
the sole purpose of serving as a target for these tests.