Use dedicated workflow step for logs coverage report
The "libraries/report-size-deltas workflow" GitHub Actions workflow prints a code coverage report to the workflow run
logs.

Previously this was done in the same step used to run the tests. Both the test run output and the coverage report will
be easier to read with a dedicated section of the logs for each, which each workflow step has. This will also make it
easier to interpret any workflow failures since the failed step would only be concerned with a single specific
operation.