Ignore failed integration test workflow job for Dependabot PRs
The integration test workflow requires the GITHUB_TOKEN token to have write permissions in order to make the report
comment.

Because the token only has read permissions when the workflow is triggered by a PR from a fork, the job is configured to
pass even though the action run step fails (it is intentionally run instead of being skipped because it can still
provide some validation even if it is expected to fail).

Currently the workflow is configured to require the step to pass when a PR is submitted from a local branch, since the
token has write permissions in that case. However, there is one exception: when the PR is submitted by Dependabot, the
token permissions are downgraded.

So the workflow must be configured to treat the PRs from Dependabot in the same manner as PRs from forks