Handle missing deltas data in sketches reports
Previously, the action's code assumed that the sketches reports artifacts would always contain size deltas data. If a
report was present that did not contain this data, the action would fail with a `KeyError` error for the missing report
key.

The most common cause of a compilation not producing size data is a compilation error. In this case the default
arduino/compile-sketches workflow configuration would not upload a sketches report workflow artifact (because the upload
is done in a subsequent step and GitHub Actions exits the job immediately when a step fails) so the
arduino/report-size-deltas action's inability to handle the resulting report format was only an issue in unusual
workflow configurations.

There is another cause of a compilation not producing size data: the boards platform was not configured to produce the
data. Previously, all known Arduino boards platforms produced at least some size data so the problem of the action's
lack of compatibility with these boards was only hypothetical. However, the "Arduino Mbed OS Portenta Boards" platform
now contains two boards which are missing the configuration to produce the size data:

- Portenta H7
- Portenta X8

This meant that the action is broken by any sketches reports produced by compiling for either of those boards.

One possible fix would be to make the arduino/compile-sketches action add deltas data objects to the sketches report
even when the compilation does not produce any data to calculate data from. But this seems like the wrong approach. So
the alternative approach was taken of making the arduino/report-size-deltas action check for the presence of the data.
If it is not present, a "N/A" is added to the report in place of the data.