Prevent unintended rendering of report content
The report is made in the form of GitHub PR comment. These comments are rendered according to the "GitHub Flavored
Markdown" markup language.

In addition to the intentional markup used to format the report, the uncontrolled strings of the report content might
contain incidental markup. Previously, this would cause the report to be rendered incorrectly.

For example, the FQBN `arduino:avr:mega:cpu=atmega2560` contains the markup for the megaphone emoji, so was rendered
incorrectly:

> arduino:avr📣cpu=atmega2560

The solution is to wrap the uncontrolled strings in the Markdown "code span" markup.

Since the entire CSV format report is already wrapped in the equivalent "code fences" markup, the introduction of this
markup into that part of the report is unnecessary.