Use latest action versions in example workflow snippets
The documentation includes some snippets of GitHub Actions workflows. In addition to arduino/report-size-deltas, these
workflows use some additional actions.

Although not essential for the purposes of documenting the usage of this project, it will be better to use the latest
version of these action dependencies to make it more likely that the readers will do the same in their own real
workflows.