Configure GITHUB_TOKEN permissions to support workflow use in private repositories
Read permissions are required in the "contents" scope in order to checkout private repositories.

Even though those permissions are not required in this project where the workflow is installed in a public repository,
this is a copy of a "template" that is intended to be applicable to any Arduino Tooling project.

A small excess in permissions in the template was chosen instead of the alternative of having to maintain separate
variants of each workflow for use in public or private repos.