Modernize spell check infrastructure
The codespell tool is used to detect commonly misspelled words in the files of the project. Infrastructure is provided for running the tool locally and a GitHub Actions workflow also runs it automatically when files are changed and periodically.

Previously the infrastructure was very outdated. It is hereby updated to the state of the art:

- codespell dependency managed by Poetry and dependabot
- Task-based to allow contributors to run the same check as done by the CI system via a standardized interface
- Use standard codespell configuration file
- Use codespell's official GitHub Actions instead of Arduino's action (deprecated in favor of the official action)