Update infrastructure for linting and add for formatting Python files
Previously linting of the Python code was done in the "libraries/report-size-deltas workflow" GitHub Actions workflow.

This system is replaced by a dedicated workflow for detecting Python code problems, leaving the
"libraries/report-size-deltas workflow" solely for running unit tests. A task is used, which will allow the contributor
to run the same check locally via a common interface as done by the CI system.

Infrastructure for formatting the Python code, and enforcing compliance with the standardized formatting style is also
added. Like the linting infrastructure, this is task-based.