Add infrastructure to check Markdown files for problems
Tasks and a GitHub Actions workflow are added for detecting problems in the project's Markdown files.

On every push and pull request that affects relevant files, and periodically, the workflow will check the repository's
Markdown files for problems:

- Use markdownlint to check for common problems and formatting.
- Use markdown-link-check to check for broken links.

The Arduino tooling Markdown style is defined by the `.markdownlint.yml` file.

In the event the repository contains externally maintained Markdown files, markdownlint can be configured to ignore them
via a `.markdownlintignore` file:
https://github.com/igorshubovych/markdownlint-cli#ignoring-files

markdown-link-check is configured via the `.markdown-link-check.json` file:
https://github.com/tcort/markdown-link-check#config-file-format