Add infrastructure for managing table of contents
The project readme contains a table of contents in order to facilitate the navigation of the documentation.

It is important that this table of contents is kept in sync with the documentation content.

A task is added for automatically generating the ToC from the content of the file.

A GitHub Actions workflow is added to check the table of contents. On every push or pull request that affects relevant
files, it will check whether the table of contents matches the file structure.