Move pytest configuration file to root of repository
The previous location worked fine when running pytest directly or via the task. It was also possible to run the tests
from the PyCharm UI after a bit of configuration. However, I had no luck with using the VS Code test explorer.

VS Code is likely to be the most common development environment of contributors so it will be important to make it easy
to use with VS Code. Moving pytest.ini file to the root of the repository allowed the VS Code test explorer to correctly
recognize it.