Add infrastructure for formatting with the Prettier tool
Prettier is a tool for formatting files of various formats. It is a standard in Arduino tooling projects.

A task is provided to allow contributors to easily format the files of the project.

A GitHub Actions workflow checks for Prettier formatting compliance on every push and pull request that affects relevant
files.