Remove invalid return type annotation from `http_request` method
The annotation was invalid, resulting in mypy error:

error: "dict" expects 2 type arguments, but 1 given  [type-arg]

The `headers` field of the dictionary returned by this function is a `http.client.HTTPMessage` class object, while the
other field values are strings. I didn't find a way to specify this in a return type annotation and the current approach
is to provide annotations on a "best effort" basis (vs making them universally mandatory) so I settled on removing the
annotation entirely.