Rework code that changed variable type through reuse
Some of the code uses practice of reusing a variable through a chain of processing its contents in some cases the
variable type was changed over the course of the processing. This caused the type checking via mypy to fail.

The code is reworked to avoid any changing of variable type.