Work with full workflow artifact data object from GitHub API
The GitHub API is used to discover the sketches report workflow artifacts. The artifact archive is then downloaded from
the URL provided by the API response. Previously, the artifact discovery methods only returned the download URL.

The API response also contains the somewhat useful information of the artifact name. Previously this information was
discarded because the name is already available indirectly from the sketches-reports-source input value. However, that
will no longer be the case once pattern matching support for the sketches-reports-source input is added. The code is
refactored to pass the full data object received from the API for the artifact between the functions. This gives direct
access to the exact artifact name.