Remove redundant path joining
The code iterates through the folder containing the sketches reports. The iteration code returns the full path to each
sketches report file. However, for some reason that path was joined to itself when it was passed to `open`. Fortunately
the pathlib module is smart enough to handle the joining of two identical absolute paths so this didn't cause a problem
but it is pointless and confusing so the redundant path joining code is removed.