Generate standard code coverage report data file in `python:test` task
The `coverage run` command used to run the Python unit tests generates a `.coverage` data file. This file can not be
consumed directly by the "codecov/codecov-action" GitHub Actions action that uploads the coverage data to Codecov. A
separate `coverage xml` command must be used to generate the appropriate file.

Previously this `coverage xml` command was ran directly in the test runner GitHub Actions workflow. However, the file is
also required by contributors running tests on their local machines in some cases (e.g., displaying coverage in the VS
Code editor via the popular "Coverage Gutters" extension, which does not support the `.coverage` file). For this reason,
it is best to move the `coverage xml` command invocation to the taskfile. The time required to generate the file is
insignificant so it is added to the `python:test` task rather than adding a separate task the contributor would be
required to run after the `python:test` task.