Move test data from test code to data files
The test_get_sketches_reports unit test makes assertions of the data the function will return given a set of input
report files. Previously, the assertion data was defined in the test code. That was a significant amount of data, which
added a lot of code to the test that is not of interest to the test code maintainer. This will become much more
problematic if additional variants of the test are added in the future to provide coverage for other conditions.

The problem is solved by moving the data to a JSON file. The test loads the data from the file when making the
assertion.