Move coverage for get_artifact failure to dedicated unit test
Previously, the test_get_artifact unit test was configured to provide coverage for the normal operation of the
get_artifact function as well as its behavior when unable to download the workflow artifact.

The code used to test the normal operation was almost completely unnecessary for testing the failure condition and the
code for adjusting the test behavior depending on which conditions were intended to be covered added significant
complexity.

Since the failure is expected to occur under normal operating conditions due to transient network outages, it is worth
continuing to provide coverage for this. So the chosen approach to providing coverage without negatively impacting the
test code is to move it to a dedicated test.