Bump action dependencies in "Upload test sketches report artifact" workflow
`actions/upload-artifact` is only bumped to v3 here because a significant change in the artifact format was introduced
in 4.0.0. It is necessary for the workflow to continue to produce test data that provides integration test coverage for
compatibility of the action with the artifacts produced by <=v3 of the action.