Generate v1 and v2 format test data artifacts
A significant change was made to the nature of GitHub Actions workflow artifacts starting from the 2.x release of the
`@actions/artifact` package, which was introduced in the 4.0.0 release of the "actions/upload-artifact" GitHub Actions
action.

It is necessary to provide integration test coverage for compatibility of the "arduino/report-size-deltas" action with
the v2 format artifacts produced by actions/upload-artifact@v4. It is also important to continue to cover compatibility
with the v1 format artifacts produced by actions/upload-artifact@v3, to support users who have not gotten around to
updating that action dependency in their sketch compilation workflows (especially since the breaking change introduced
by actions/upload-artifact@v4 will make it necessary to adjust the configuration of the workflow).

The chosen approach is to configure the "Upload test sketches report artifact" workflow to use
actions/upload-artifact@v3 when upload one of the test data artifacts and actions/upload-artifact@v4 when uploading the
other. Even though the produced test data would not occur under real world usage, it is the cleanest way (the
alternative being to create and maintain separate copies of the system implemented in the `report-target-pr` branch for
each major version of the actions/upload-artifact action) to provide effective integration test coverage.