Restrict HTTP request error handling to appropriate exception
The "arduino/report-size-deltas" action relies on the GitHub REST API. It makes HTTP requests to various API endpoints
during each workflow run.

It is expected that these requests might occasionally produce an exception due to transient network outages. For this
reason, the action has special handling of exceptions. Previously, that exception handling code was configured to be
triggered on any exception, even though it is specific to HTTP request errors. It is more appropriate to target it only
to the specific urllib.error.HTTPError exception that results from an HTTP request error, leaving any other types of
unexpected exceptions unhandled.