Increase visibility of error message when action fails due to HTTP request error
The "arduino/report-size-deltas" action relies on the GitHub REST API. It makes HTTP requests to various API endpoints
during each workflow run.

These requests might produce an exception due to transient network outages, misconfiguration of the workflow, or a
problem with the action. This will cause the action's step in the workflow run to fail. The maintainer of the repository
in which the action is used may then need to investigate the cause of the failure.

Previously, the cause of the failure could only be determined by evaluating the workflow run logs, which are somewhat
unfriendly due to containing the Python stack trace produced by the HTTP request exception. The significant error
message is hereby made visible as an annotation on the workflow run summary page through the use of an "error" workflow
command.