Configure pytest to capture test output
By default, pytest captures all output from the code during test execution. For unknown reasons, the creator of the
original pytest configuration file which was reused for this project decided to disable that behavior. This resulted in
any output from the code under test being shown in the test logs, making them very difficult to read and interpret.

That configuration is hereby removed. After this change, script output from successful test runs will not be shown in
the test logs. Since the output might be useful for troubleshooting test failures, it will be shown when a test fails,
in a special section of the logs.