Pin GitHub REST API version in requests
Previously, the requests would use whatever version of the API is current, exposing the action to immediate breakage if
GitHub released a new version of the REST API with relevant changes.

GitHub provides the previous API for a minimum of 24 months after a new version is released, so pinning the REST API
version used by the action via the `X-GitHub-Api-Version` HTTP header improves the stability of the action and allows
the project maintainer to make a controlled migration to the new API version.