Use latest versions of action dependencies in example workflows
Some GitHub Actions workflows are provided in the readme to demonstrate the use of the action. These workflows have
dependencies on the "actions/upload-artifact" and "actions/download-artifact" actions.

Previously, these dependencies were pinned to their 3.x major version series. A 4.x major version series of the actions
was recently started. It is best practices to use the latest version of actions, so the documentation should demonstrate
that.

A significant breaking change was introduced in the 4.0.0 release of "actions/upload-artifact": uploading multiple times
to the same artifact is no longer allowed. Previously the "Workflow triggered by `pull_request` event" example workflow
used a single artifact for the sketches report file produced by the matrix job for each of the boards the sketches were
compiled for. It was necessary to adjust the configuration of the workflow to use a separate workflow for each of the
sketches report files in order to accommodate the bump fo the "actions/upload-artifact" action dependency.