Add demonstration workflows for multi-board coverage
The readme contains a set of workflows demonstrating the use of the action in the use case where the report workflow is
triggered by a `schedule` event.

These workflows demonstrate the most simple use case, where sketches are only being compiled for a single board.
Although this provides a good introduction into using the action, the action user will often want the system to provide
coverage for multiple boards. The configuration of the workflows become significantly more complex, especially now that
it is necessary to upload each sketches report to a separate artifact. It will be beneficial to also demonstrate that
advanced usage.

The demonstration consists of a significant amount of content so putting it in the readme would harm the readability and
approachability of that document. For this reason, it is published in a separate document which is linked to from the
readme. This "FAQ" document will serve as a container for the addition of more supplemental documentation in the future.