Add simple demonstration workflow for `pull_request`-triggered workflow
A demonstration workflow is provided in the readme to show how to use the action in the use case where the report
workflow is triggered by `pull_request` event.

Previously, this workflow demonstrated the more advanced configuration that is required in order to compile for multiple
boards. Although this demonstration will be useful to the action users, it might be overwhelming as an introduction to
the action usage. A demonstration that only compiles for a single board can be far more simple. Even if the user
requires coverage of multiple boards in their application, this will give them a basic understanding of the action
usage.

The advanced demonstration workflow is hereby moved to the FAQ document, replaced in the readme by a simple
demonstration workflow. A link from the readme to the advanced usage demonstration is provided.