Remove non-functional bold tags from Mermaid code
In order to make it easier for the reader to understand the function of complementary example workflows provided in the FAQ, a chart is used to explain the flow of sketches report data between the steps and jobs in the sketch compilation workflow and between the sketch compilation workflow and the deltas report workflow.

The Mermaid diagramming language is used to produce this flowchart (GitHub automatically renders Mermaid code in code blocks of Markdown files).

The text in charts can be styled and formatted using HTML tags in the Mermaid code. Previously, bold tags were used on the subgraph titles in order to improve the readability of the chart. Since the time the chart was created, something changed in GitHub's Markdown/Mermaid rendering system that causes these tags to no longer cause the intended text styling. They now are simply printed as text, meaning they now have a harmful effect on the readability. For this reason, the tags are hereby removed. No alternative is available. The line break tags continue to work as expected (though unfortunately the text is now left aligned instead of being centered as before) so they are retained.