Bump project Node.js version to 22
npm package-based tools are used for various development and maintenance operations for this project. In order to ensure
these operations work as expected, the project is configured for a specific major version of Node.js to be used by the
infrastructure systems and collaborators.

The standard Node.js version for Arduino Tooling projects is now 22.